Moved helped functions to spin.c and changed naming convention
